Alternative approach. Using ruby's test/unit
