--- test/integration/tests/0-noop.rb
@@ -1,4 +1,0 @@
-# this test does nothing
-# it's, in a way, a test of the integration testing framework
-
-puts 'executing noop test'
